fix: reduce complexity by splitting transfer into parts
