+++ manta-accounting/src/transfer/mod.rs
@@ -31,18 +31,14 @@
 
 use crate::asset::{Asset, AssetId, AssetValue};
 use alloc::vec::Vec;
-use core::{fmt::Debug, hash::Hash, iter, marker::PhantomData, ops::Deref};
+use core::{fmt::Debug, hash::Hash, marker::PhantomData, ops::Deref};
 use manta_crypto::{
-    accumulator::{Accumulator, AssertValidVerification, MembershipProof, Model},
+    accumulator::{AssertValidVerification, MembershipProof, Model},
     constraint::{
         self, Add, Allocate, Allocator, AssertEq, Bool, Constant, Derived, ProofSystem,
         ProofSystemInput, Public, Secret, Variable,
     },
-    encryption::{
-        self,
-        hybrid::{self, Hybrid},
-        Encrypt, EncryptedMessage,
-    },
+    encryption::{self, hybrid::Hybrid, EncryptedMessage},
     key::{self, agreement::Derive},
     rand::{CryptoRng, RngCore, Sample},
 };
@@ -51,6 +47,9 @@ use manta_util::SizeLimit;
 #[cfg(feature = "serde")]
 use manta_util::serde::{Deserialize, Serialize};
 
+mod receiver;
+mod sender;
+
 pub mod batch;
 pub mod canonical;
 
@@ -58,8 +57,9 @@ pub mod canonical;
 #[cfg_attr(doc_cfg, doc(cfg(feature = "test")))]
 pub mod test;
 
-#[doc(inline)]
 pub use canonical::Shape;
+pub use receiver::*;
+pub use sender::*;
 
 /// Returns `true` if the [`Transfer`] with this shape would have public participants.
 #[inline]
@@ -752,536 +752,6 @@ where
     }
 }
 
-/// Pre-Sender
-pub struct PreSender<C>
-where
-    C: Configuration,
-{
-    /// Secret Spend Key
-    secret_spend_key: SecretKey<C>,
-
-    /// Ephemeral Secret Key
-    ephemeral_secret_key: SecretKey<C>,
-
-    /// Asset
-    asset: Asset,
-
-    /// Unspent Transaction Output
-    utxo: Utxo<C>,
-
-    /// Void Number
-    void_number: VoidNumber<C>,
-}
-
-impl<C> PreSender<C>
-where
-    C: Configuration,
-{
-    /// Builds a new [`PreSender`] from `ephemeral_secret_key` to claim `asset` with
-    /// `secret_spend_key`.
-    #[inline]
-    pub fn new(
-        parameters: &Parameters<C>,
-        secret_spend_key: SecretKey<C>,
-        ephemeral_secret_key: SecretKey<C>,
-        asset: Asset,
-    ) -> Self {
-        let utxo = parameters.utxo(
-            &ephemeral_secret_key,
-            &parameters.derive(&secret_spend_key),
-            &asset,
-        );
-        Self {
-            void_number: parameters.void_number(&secret_spend_key, &utxo),
-            secret_spend_key,
-            ephemeral_secret_key,
-            asset,
-            utxo,
-        }
-    }
-
-    /// Inserts the [`Utxo`] corresponding to `self` into the `utxo_accumulator` with the intention
-    /// of returning a proof later by a call to [`get_proof`](Self::get_proof).
-    #[inline]
-    pub fn insert_utxo<A>(&self, utxo_accumulator: &mut A) -> bool
-    where
-        A: Accumulator<Item = Utxo<C>, Model = C::UtxoAccumulatorModel>,
-    {
-        utxo_accumulator.insert(&self.utxo)
-    }
-
-    /// Requests the membership proof of the [`Utxo`] corresponding to `self` from
-    /// `utxo_accumulator` to prepare the conversion from `self` into a [`Sender`].
-    #[inline]
-    pub fn get_proof<A>(&self, utxo_accumulator: &A) -> Option<SenderProof<C>>
-    where
-        A: Accumulator<Item = Utxo<C>, Model = C::UtxoAccumulatorModel>,
-    {
-        Some(SenderProof {
-            utxo_membership_proof: utxo_accumulator.prove(&self.utxo)?,
-        })
-    }
-
-    /// Converts `self` into a [`Sender`] by attaching `proof` to it.
-    #[inline]
-    pub fn upgrade(self, proof: SenderProof<C>) -> Sender<C> {
-        Sender {
-            secret_spend_key: self.secret_spend_key,
-            ephemeral_secret_key: self.ephemeral_secret_key,
-            asset: self.asset,
-            utxo: self.utxo,
-            utxo_membership_proof: proof.utxo_membership_proof,
-            void_number: self.void_number,
-        }
-    }
-
-    /// Tries to convert `self` into a [`Sender`] by getting a proof from `utxo_accumulator`.
-    #[inline]
-    pub fn try_upgrade<A>(self, utxo_accumulator: &A) -> Option<Sender<C>>
-    where
-        A: Accumulator<Item = Utxo<C>, Model = C::UtxoAccumulatorModel>,
-    {
-        Some(self.get_proof(utxo_accumulator)?.upgrade(self))
-    }
-
-    /// Inserts the [`Utxo`] corresponding to `self` into the `utxo_accumulator` and upgrades to a
-    /// full [`Sender`] if the insertion succeeded.
-    #[inline]
-    pub fn insert_and_upgrade<A>(self, utxo_accumulator: &mut A) -> Option<Sender<C>>
-    where
-        A: Accumulator<Item = Utxo<C>, Model = C::UtxoAccumulatorModel>,
-    {
-        if self.insert_utxo(utxo_accumulator) {
-            self.try_upgrade(utxo_accumulator)
-        } else {
-            None
-        }
-    }
-
-    /// Returns `true` whenever `self.utxo` and `rhs.utxo` can be inserted in any order into the
-    /// `utxo_accumulator`.
-    #[inline]
-    pub fn is_independent_from<A>(&self, rhs: &Self, utxo_accumulator: &A) -> bool
-    where
-        A: Accumulator<Item = Utxo<C>, Model = C::UtxoAccumulatorModel>,
-    {
-        utxo_accumulator.are_independent(&self.utxo, &rhs.utxo)
-    }
-}
-
-/// Sender Proof
-///
-/// This `struct` is created by the [`get_proof`](PreSender::get_proof) method on [`PreSender`].
-/// See its documentation for more.
-pub struct SenderProof<C>
-where
-    C: Configuration,
-{
-    /// UTXO Membership Proof
-    utxo_membership_proof: UtxoMembershipProof<C>,
-}
-
-impl<C> SenderProof<C>
-where
-    C: Configuration,
-{
-    /// Upgrades the `pre_sender` to a [`Sender`] by attaching `self` to it.
-    #[inline]
-    pub fn upgrade(self, pre_sender: PreSender<C>) -> Sender<C> {
-        pre_sender.upgrade(self)
-    }
-}
-
-/// Sender
-pub struct Sender<C>
-where
-    C: Configuration,
-{
-    /// Secret Spend Key
-    secret_spend_key: SecretKey<C>,
-
-    /// Ephemeral Secret Key
-    ephemeral_secret_key: SecretKey<C>,
-
-    /// Asset
-    asset: Asset,
-
-    /// Unspent Transaction Output
-    utxo: Utxo<C>,
-
-    /// UTXO Membership Proof
-    utxo_membership_proof: UtxoMembershipProof<C>,
-
-    /// Void Number
-    void_number: VoidNumber<C>,
-}
-
-impl<C> Sender<C>
-where
-    C: Configuration,
-{
-    /// Returns the asset value sent by `self` in the transaction.
-    #[inline]
-    pub fn asset_value(&self) -> AssetValue {
-        self.asset.value
-    }
-
-    /// Returns `true` whenever `self.utxo` and `rhs.utxo` can be inserted in any order into the
-    /// `utxo_accumulator`.
-    #[inline]
-    pub fn is_independent_from<A>(&self, rhs: &Self, utxo_accumulator: &A) -> bool
-    where
-        A: Accumulator<Item = Utxo<C>, Model = C::UtxoAccumulatorModel>,
-    {
-        utxo_accumulator.are_independent(&self.utxo, &rhs.utxo)
-    }
-
-    /// Reverts `self` back into a [`PreSender`].
-    ///
-    /// This method should be called if the [`Utxo`] membership proof attached to `self` was deemed
-    /// invalid or had expired.
-    #[inline]
-    pub fn downgrade(self) -> PreSender<C> {
-        PreSender {
-            secret_spend_key: self.secret_spend_key,
-            ephemeral_secret_key: self.ephemeral_secret_key,
-            asset: self.asset,
-            utxo: self.utxo,
-            void_number: self.void_number,
-        }
-    }
-
-    /// Extracts the ledger posting data from `self`.
-    #[inline]
-    pub fn into_post(self) -> SenderPost<C> {
-        SenderPost {
-            utxo_accumulator_output: self.utxo_membership_proof.into_output(),
-            void_number: self.void_number,
-        }
-    }
-
-    /// Extends proof public input with `self`.
-    #[inline]
-    pub fn extend_input(&self, input: &mut ProofInput<C>) {
-        C::ProofSystem::extend(input, self.utxo_membership_proof.output());
-        C::ProofSystem::extend(input, &self.void_number);
-    }
-}
-
-/// Sender Variable
-pub struct SenderVar<C>
-where
-    C: Configuration,
-{
-    /// Secret Spend Key
-    secret_spend_key: SecretKeyVar<C>,
-
-    /// Ephemeral Secret Key
-    ephemeral_secret_key: SecretKeyVar<C>,
-
-    /// Asset
-    asset: AssetVar<C>,
-
-    /// UTXO Membership Proof
-    utxo_membership_proof: UtxoMembershipProofVar<C>,
-
-    /// Void Number
-    void_number: VoidNumberVar<C>,
-}
-
-impl<C> SenderVar<C>
-where
-    C: Configuration,
-{
-    /// Returns the asset for `self`, checking if `self` is well-formed.
-    #[inline]
-    pub fn get_well_formed_asset(
-        self,
-        parameters: &FullParametersVar<C>,
-        compiler: &mut C::Compiler,
-    ) -> AssetVar<C> {
-        let public_spend_key = parameters.derive(&self.secret_spend_key, compiler);
-        let utxo = parameters.utxo(
-            &self.ephemeral_secret_key,
-            &public_spend_key,
-            &self.asset,
-            compiler,
-        );
-        self.utxo_membership_proof.assert_valid(
-            &parameters.utxo_accumulator_model,
-            &utxo,
-            compiler,
-        );
-        let void_number = parameters.void_number(&self.secret_spend_key, &utxo, compiler);
-        compiler.assert_eq(&self.void_number, &void_number);
-        self.asset
-    }
-}
-
-impl<C> Variable<Derived, C::Compiler> for SenderVar<C>
-where
-    C: Configuration,
-{
-    type Type = Sender<C>;
-
-    #[inline]
-    fn new_known(this: &Self::Type, compiler: &mut C::Compiler) -> Self {
-        Self {
-            secret_spend_key: this.secret_spend_key.as_known(compiler),
-            ephemeral_secret_key: this.ephemeral_secret_key.as_known(compiler),
-            asset: this.asset.as_known(compiler),
-            utxo_membership_proof: this.utxo_membership_proof.as_known(compiler),
-            void_number: this.void_number.as_known(compiler),
-        }
-    }
-
-    #[inline]
-    fn new_unknown(compiler: &mut C::Compiler) -> Self {
-        Self {
-            secret_spend_key: compiler.allocate_unknown(),
-            ephemeral_secret_key: compiler.allocate_unknown(),
-            asset: compiler.allocate_unknown(),
-            utxo_membership_proof: compiler.allocate_unknown(),
-            void_number: compiler.allocate_unknown(),
-        }
-    }
-}
-
-/// Sender Ledger
-///
-/// This is the validation trait for ensuring that a particular instance of [`Sender`] is valid
-/// according to the ledger state. These methods are the minimum required for a ledger which accepts
-/// the [`Sender`] abstraction.
-pub trait SenderLedger<C>
-where
-    C: Configuration,
-{
-    /// Valid [`VoidNumber`] Posting Key
-    ///
-    /// # Safety
-    ///
-    /// This type must be some wrapper around [`VoidNumber`] which can only be constructed by this
-    /// implementation of [`SenderLedger`]. This is to prevent that [`spend`](Self::spend) is
-    /// called before [`is_unspent`](Self::is_unspent) and
-    /// [`has_matching_utxo_accumulator_output`](Self::has_matching_utxo_accumulator_output).
-    type ValidVoidNumber: AsRef<VoidNumber<C>>;
-
-    /// Valid UTXO Accumulator Output Posting Key
-    ///
-    /// # Safety
-    ///
-    /// This type must be some wrapper around [`S::Output`] which can only be constructed by this
-    /// implementation of [`SenderLedger`]. This is to prevent that [`spend`](Self::spend) is
-    /// called before [`is_unspent`](Self::is_unspent) and
-    /// [`has_matching_utxo_accumulator_output`](Self::has_matching_utxo_accumulator_output).
-    ///
-    /// [`S::Output`]: Model::Output
-    type ValidUtxoAccumulatorOutput: AsRef<UtxoAccumulatorOutput<C>>;
-
-    /// Super Posting Key
-    ///
-    /// Type that allows super-traits of [`SenderLedger`] to customize posting key behavior.
-    type SuperPostingKey: Copy;
-
-    /// Checks if the ledger already contains the `void_number` in its set of void numbers.
-    ///
-    /// Existence of such a void number could indicate a possible double-spend.
-    fn is_unspent(&self, void_number: VoidNumber<C>) -> Option<Self::ValidVoidNumber>;
-
-    /// Checks if `output` matches the current accumulated value of the UTXO accumulator that is
-    /// stored on the ledger.
-    ///
-    /// Failure to match the ledger state means that the sender was constructed under an invalid or
-    /// older state of the ledger.
-    fn has_matching_utxo_accumulator_output(
-        &self,
-        output: UtxoAccumulatorOutput<C>,
-    ) -> Option<Self::ValidUtxoAccumulatorOutput>;
-
-    /// Posts the `void_number` to the ledger, spending the asset.
-    ///
-    /// # Safety
-    ///
-    /// This method can only be called once we check that `void_number` is not already stored on
-    /// the ledger. See [`is_unspent`](Self::is_unspent) for more.
-    ///
-    /// # Implementation Note
-    ///
-    /// This method, by defualt, calls the [`spend_all`] method on an iterator of length one
-    /// containing `(utxo, encrypted_note)`. Either [`spend`] or [`spend_all`] can be implemented
-    /// depending on which is more efficient.
-    ///
-    /// [`spend`]: Self::spend
-    /// [`spend_all`]: Self::spend_all
-    #[inline]
-    fn spend(
-        &mut self,
-        utxo_accumulator_output: Self::ValidUtxoAccumulatorOutput,
-        void_number: Self::ValidVoidNumber,
-        super_key: &Self::SuperPostingKey,
-    ) {
-        self.spend_all(
-            iter::once((utxo_accumulator_output, void_number)),
-            super_key,
-        )
-    }
-
-    /// Posts all of the [`VoidNumber`] to the ledger, spending the assets.
-    ///
-    /// # Safety
-    ///
-    /// This method can only be called once we check that all the [`VoidNumber`] are not already
-    /// stored on the ledger. See [`is_unspent`](Self::is_unspent) for more.
-    ///
-    /// # Implementation Note
-    ///
-    /// This method is an optimization path for multiple calls to [`spend`] and by default just
-    /// iterates over `iter` calling [`spend`] on each item returned. Either [`spend`] or
-    /// [`spend_all`] can be implemented depending on which is more efficient.
-    ///
-    /// [`spend`]: Self::spend
-    /// [`spend_all`]: Self::spend_all
-    #[inline]
-    fn spend_all<I>(&mut self, iter: I, super_key: &Self::SuperPostingKey)
-    where
-        I: IntoIterator<Item = (Self::ValidUtxoAccumulatorOutput, Self::ValidVoidNumber)>,
-    {
-        for (utxo_accumulator_output, void_number) in iter {
-            self.spend(utxo_accumulator_output, void_number, super_key)
-        }
-    }
-}
-
-/// Sender Post Error
-#[cfg_attr(
-    feature = "serde",
-    derive(Deserialize, Serialize),
-    serde(crate = "manta_util::serde", deny_unknown_fields)
-)]
-#[derive(Clone, Copy, Debug, Eq, Hash, PartialEq)]
-pub enum SenderPostError {
-    /// Asset Spent Error
-    ///
-    /// The asset has already been spent.
-    AssetSpent,
-
-    /// Invalid UTXO Accumulator Output Error
-    ///
-    /// The sender was not constructed under the current state of the UTXO accumulator.
-    InvalidUtxoAccumulatorOutput,
-}
-
-/// Sender Post
-///
-/// This `struct` represents the public data required to verify that a particular instance of a
-/// [`Sender`] should be valid according to the [`SenderLedger`]. The rest of the information
-/// required to verify a [`Transfer`] is stored in the [`TransferPost`] which includes the [`Proof`]
-/// of validity.
-#[cfg_attr(
-    feature = "serde",
-    derive(Deserialize, Serialize),
-    serde(
-        bound(
-            deserialize = "UtxoAccumulatorOutput<C>: Deserialize<'de>, VoidNumber<C>: Deserialize<'de>",
-            serialize = "UtxoAccumulatorOutput<C>: Serialize, VoidNumber<C>: Serialize",
-        ),
-        crate = "manta_util::serde",
-        deny_unknown_fields
-    )
-)]
-#[derive(derivative::Derivative)]
-#[derivative(
-    Clone(bound = "UtxoAccumulatorOutput<C>: Clone, VoidNumber<C>: Clone"),
-    Copy(bound = "UtxoAccumulatorOutput<C>: Copy, VoidNumber<C>: Copy"),
-    Debug(bound = "UtxoAccumulatorOutput<C>: Debug, VoidNumber<C>: Debug"),
-    Eq(bound = "UtxoAccumulatorOutput<C>: Eq, VoidNumber<C>: Eq"),
-    Hash(bound = "UtxoAccumulatorOutput<C>: Hash, VoidNumber<C>: Hash"),
-    PartialEq(bound = "UtxoAccumulatorOutput<C>: PartialEq, VoidNumber<C>: PartialEq")
-)]
-pub struct SenderPost<C>
-where
-    C: Configuration,
-{
-    /// UTXO Accumulator Output
-    pub utxo_accumulator_output: UtxoAccumulatorOutput<C>,
-
-    /// Void Number
-    pub void_number: VoidNumber<C>,
-}
-
-impl<C> SenderPost<C>
-where
-    C: Configuration,
-{
-    /// Extends proof public input with `self`.
-    #[inline]
-    pub fn extend_input(&self, input: &mut ProofInput<C>) {
-        C::ProofSystem::extend(input, &self.utxo_accumulator_output);
-        C::ProofSystem::extend(input, &self.void_number);
-    }
-
-    /// Validates `self` on the sender `ledger`.
-    #[inline]
-    pub fn validate<L>(self, ledger: &L) -> Result<SenderPostingKey<C, L>, SenderPostError>
-    where
-        L: SenderLedger<C>,
-    {
-        Ok(SenderPostingKey {
-            utxo_accumulator_output: ledger
-                .has_matching_utxo_accumulator_output(self.utxo_accumulator_output)
-                .ok_or(SenderPostError::InvalidUtxoAccumulatorOutput)?,
-            void_number: ledger
-                .is_unspent(self.void_number)
-                .ok_or(SenderPostError::AssetSpent)?,
-        })
-    }
-}
-
-/// Sender Posting Key
-pub struct SenderPostingKey<C, L>
-where
-    C: Configuration,
-    L: SenderLedger<C> + ?Sized,
-{
-    /// UTXO Accumulator Output Posting Key
-    utxo_accumulator_output: L::ValidUtxoAccumulatorOutput,
-
-    /// Void Number Posting Key
-    void_number: L::ValidVoidNumber,
-}
-
-impl<C, L> SenderPostingKey<C, L>
-where
-    C: Configuration,
-    L: SenderLedger<C> + ?Sized,
-{
-    /// Extends proof public input with `self`.
-    #[inline]
-    pub fn extend_input(&self, input: &mut ProofInput<C>) {
-        C::ProofSystem::extend(input, self.utxo_accumulator_output.as_ref());
-        C::ProofSystem::extend(input, self.void_number.as_ref());
-    }
-
-    /// Posts `self` to the sender `ledger`.
-    #[inline]
-    pub fn post(self, super_key: &L::SuperPostingKey, ledger: &mut L) {
-        ledger.spend(self.utxo_accumulator_output, self.void_number, super_key);
-    }
-
-    /// Posts all of the [`SenderPostingKey`] in `iter` to the sender `ledger`.
-    #[inline]
-    pub fn post_all<I>(iter: I, super_key: &L::SuperPostingKey, ledger: &mut L)
-    where
-        I: IntoIterator<Item = Self>,
-    {
-        ledger.spend_all(
-            iter.into_iter()
-                .map(move |k| (k.utxo_accumulator_output, k.void_number)),
-            super_key,
-        )
-    }
-}
-
 /// Note
 #[derive(derivative::Derivative)]
 #[derivative(
@@ -1340,368 +810,6 @@ where
     C: Configuration,
 {
     const SIZE: usize = SecretKey::<C>::SIZE + Asset::SIZE;
-}
-
-/// Receiver
-pub struct Receiver<C>
-where
-    C: Configuration,
-{
-    /// Public Spend Key
-    public_spend_key: PublicKey<C>,
-
-    /// Ephemeral Secret Spend Key
-    ephemeral_secret_key: SecretKey<C>,
-
-    /// Asset
-    asset: Asset,
-
-    /// Unspent Transaction Output
-    utxo: Utxo<C>,
-
-    /// Encrypted Note
-    encrypted_note: EncryptedNote<C>,
-}
-
-impl<C> Receiver<C>
-where
-    C: Configuration,
-{
-    /// Build a new [`Receiver`] from `ephemeral_secret_key`, to send `asset` to the owners of
-    /// `public_spend_key` and `public_view_key`.
-    #[inline]
-    pub fn new(
-        parameters: &Parameters<C>,
-        public_spend_key: PublicKey<C>,
-        public_view_key: PublicKey<C>,
-        ephemeral_secret_key: SecretKey<C>,
-        asset: Asset,
-    ) -> Self {
-        let randomness = hybrid::Randomness::from_key(ephemeral_secret_key);
-        Self {
-            utxo: parameters.utxo(&randomness.ephemeral_secret_key, &public_spend_key, &asset),
-            encrypted_note: parameters.note_encryption_scheme.encrypt_into(
-                &public_view_key,
-                &randomness,
-                (),
-                &Note::new(randomness.ephemeral_secret_key.clone(), asset),
-                &mut (),
-            ),
-            ephemeral_secret_key: randomness.ephemeral_secret_key,
-            public_spend_key,
-            asset,
-        }
-    }
-
-    /// Returns the ephemeral public key associated to `self`.
-    #[inline]
-    pub fn ephemeral_public_key(&self) -> &PublicKey<C> {
-        self.encrypted_note.ephemeral_public_key()
-    }
-
-    /// Returns `true` whenever `self.utxo` and `rhs.utxo` can be inserted in any order into the
-    /// `utxo_accumulator`.
-    #[inline]
-    pub fn is_independent_from<A>(&self, rhs: &Self, utxo_accumulator: &A) -> bool
-    where
-        A: Accumulator<Item = Utxo<C>, Model = C::UtxoAccumulatorModel>,
-    {
-        utxo_accumulator.are_independent(&self.utxo, &rhs.utxo)
-    }
-
-    /// Extracts the ledger posting data from `self`.
-    #[inline]
-    pub fn into_post(self) -> ReceiverPost<C> {
-        ReceiverPost {
-            utxo: self.utxo,
-            encrypted_note: self.encrypted_note,
-        }
-    }
-
-    /// Extends proof public input with `self`.
-    #[inline]
-    pub fn extend_input(&self, input: &mut ProofInput<C>) {
-        C::ProofSystem::extend(input, &self.utxo);
-    }
-}
-
-/// Receiver Variable
-pub struct ReceiverVar<C>
-where
-    C: Configuration,
-{
-    /// Ephemeral Secret Spend Key
-    ephemeral_secret_key: SecretKeyVar<C>,
-
-    /// Public Spend Key
-    public_spend_key: PublicKeyVar<C>,
-
-    /// Asset
-    asset: AssetVar<C>,
-
-    /// Unspent Transaction Output
-    utxo: UtxoVar<C>,
-}
-
-impl<C> ReceiverVar<C>
-where
-    C: Configuration,
-{
-    /// Returns the asset for `self`, checking if `self` is well-formed.
-    #[inline]
-    pub fn get_well_formed_asset(
-        self,
-        parameters: &FullParametersVar<C>,
-        compiler: &mut C::Compiler,
-    ) -> AssetVar<C> {
-        let utxo = parameters.utxo(
-            &self.ephemeral_secret_key,
-            &self.public_spend_key,
-            &self.asset,
-            compiler,
-        );
-        compiler.assert_eq(&self.utxo, &utxo);
-        self.asset
-    }
-}
-
-impl<C> Variable<Derived, C::Compiler> for ReceiverVar<C>
-where
-    C: Configuration,
-{
-    type Type = Receiver<C>;
-
-    #[inline]
-    fn new_known(this: &Self::Type, compiler: &mut C::Compiler) -> Self {
-        Self {
-            ephemeral_secret_key: this.ephemeral_secret_key.as_known(compiler),
-            public_spend_key: this.public_spend_key.as_known(compiler),
-            asset: this.asset.as_known(compiler),
-            utxo: this.utxo.as_known::<Public, _>(compiler),
-        }
-    }
-
-    #[inline]
-    fn new_unknown(compiler: &mut C::Compiler) -> Self {
-        Self {
-            ephemeral_secret_key: compiler.allocate_unknown(),
-            public_spend_key: compiler.allocate_unknown(),
-            asset: compiler.allocate_unknown(),
-            utxo: compiler.allocate_unknown::<Public, _>(),
-        }
-    }
-}
-
-/// Receiver Ledger
-///
-/// This is the validation trait for ensuring that a particular instance of [`Receiver`] is valid
-/// according to the ledger state. These methods are the minimum required for a ledger which accepts
-/// the [`Receiver`] abstraction.
-pub trait ReceiverLedger<C>
-where
-    C: Configuration,
-{
-    /// Valid [`Utxo`] Posting Key
-    ///
-    /// # Safety
-    ///
-    /// This type must be some wrapper around [`Utxo`] which can only be constructed by this
-    /// implementation of [`ReceiverLedger`]. This is to prevent that [`register`](Self::register)
-    /// is called before [`is_not_registered`](Self::is_not_registered).
-    type ValidUtxo: AsRef<Utxo<C>>;
-
-    /// Super Posting Key
-    ///
-    /// Type that allows super-traits of [`ReceiverLedger`] to customize posting key behavior.
-    type SuperPostingKey: Copy;
-
-    /// Checks if the ledger already contains the `utxo` in its set of UTXOs.
-    ///
-    /// Existence of such a UTXO could indicate a possible double-spend.
-    fn is_not_registered(&self, utxo: Utxo<C>) -> Option<Self::ValidUtxo>;
-
-    /// Posts the `utxo` and `encrypted_note` to the ledger, registering the asset.
-    ///
-    /// # Safety
-    ///
-    /// This method can only be called once we check that `utxo` is not already stored on the
-    /// ledger. See [`is_not_registered`](Self::is_not_registered) for more.
-    ///
-    /// # Implementation Note
-    ///
-    /// This method, by default, calls the [`register_all`] method on an iterator of length one
-    /// containing `(utxo, encrypted_note)`. Either [`register`] or [`register_all`] can be
-    /// implemented depending on which is more efficient.
-    ///
-    /// [`register`]: Self::register
-    /// [`register_all`]: Self::register_all
-    #[inline]
-    fn register(
-        &mut self,
-        utxo: Self::ValidUtxo,
-        encrypted_note: EncryptedNote<C>,
-        super_key: &Self::SuperPostingKey,
-    ) {
-        self.register_all(iter::once((utxo, encrypted_note)), super_key)
-    }
-
-    /// Posts all of the [`Utxo`] and [`EncryptedNote`] to the ledger, registering the assets.
-    ///
-    /// # Safety
-    ///
-    /// This method can only be called once we check that all the [`Utxo`] and [`EncryptedNote`] are
-    /// not already stored on the ledger. See [`is_not_registered`](Self::is_not_registered) for
-    /// more.
-    ///
-    /// # Implementation Note
-    ///
-    /// This method is an optimization path for multiple calls to [`register`] and by default just
-    /// iterates over `iter` calling [`register`] on each item returned. Either [`register`] or
-    /// [`register_all`] can be implemented depending on which is more efficient.
-    ///
-    /// [`register`]: Self::register
-    /// [`register_all`]: Self::register_all
-    #[inline]
-    fn register_all<I>(&mut self, iter: I, super_key: &Self::SuperPostingKey)
-    where
-        I: IntoIterator<Item = (Self::ValidUtxo, EncryptedNote<C>)>,
-    {
-        for (utxo, encrypted_note) in iter {
-            self.register(utxo, encrypted_note, super_key)
-        }
-    }
-}
-
-/// Receiver Post Error
-#[cfg_attr(
-    feature = "serde",
-    derive(Deserialize, Serialize),
-    serde(crate = "manta_util::serde", deny_unknown_fields)
-)]
-#[derive(Clone, Copy, Debug, Eq, Hash, PartialEq)]
-pub enum ReceiverPostError {
-    /// Asset Registered Error
-    ///
-    /// The asset has already been registered with the ledger.
-    AssetRegistered,
-}
-
-/// Receiver Post
-///
-/// This `struct` represents the public data required to verify that a particular instance of a
-/// [`Receiver`] should be valid according to the [`ReceiverLedger`]. The rest of the information
-/// required to verify a [`Transfer`] is stored in the [`TransferPost`] which includes the [`Proof`]
-/// of validity.
-#[cfg_attr(
-    feature = "serde",
-    derive(Deserialize, Serialize),
-    serde(
-        bound(
-            deserialize = "Utxo<C>: Deserialize<'de>, EncryptedNote<C>: Deserialize<'de>",
-            serialize = "Utxo<C>: Serialize, EncryptedNote<C>: Serialize",
-        ),
-        crate = "manta_util::serde",
-        deny_unknown_fields
-    )
-)]
-#[derive(derivative::Derivative)]
-#[derivative(
-    Clone(bound = "Utxo<C>: Clone, EncryptedNote<C>: Clone"),
-    Copy(bound = "Utxo<C>: Copy, EncryptedNote<C>: Copy"),
-    Debug(bound = "Utxo<C>: Debug, EncryptedNote<C>: Debug"),
-    Eq(bound = "Utxo<C>: Eq, EncryptedNote<C>: Eq"),
-    Hash(bound = "Utxo<C>: Hash, EncryptedNote<C>: Hash"),
-    PartialEq(bound = "Utxo<C>: PartialEq, EncryptedNote<C>: PartialEq")
-)]
-pub struct ReceiverPost<C>
-where
-    C: Configuration,
-{
-    /// Unspent Transaction Output
-    pub utxo: Utxo<C>,
-
-    /// Encrypted Note
-    pub encrypted_note: EncryptedNote<C>,
-}
-
-impl<C> ReceiverPost<C>
-where
-    C: Configuration,
-{
-    /// Returns the ephemeral public key associated to `self`.
-    #[inline]
-    pub fn ephemeral_public_key(&self) -> &PublicKey<C> {
-        self.encrypted_note.ephemeral_public_key()
-    }
-
-    /// Extends proof public input with `self`.
-    #[inline]
-    pub fn extend_input(&self, input: &mut ProofInput<C>) {
-        C::ProofSystem::extend(input, &self.utxo);
-    }
-
-    /// Validates `self` on the receiver `ledger`.
-    #[inline]
-    pub fn validate<L>(self, ledger: &L) -> Result<ReceiverPostingKey<C, L>, ReceiverPostError>
-    where
-        L: ReceiverLedger<C>,
-    {
-        Ok(ReceiverPostingKey {
-            utxo: ledger
-                .is_not_registered(self.utxo)
-                .ok_or(ReceiverPostError::AssetRegistered)?,
-            encrypted_note: self.encrypted_note,
-        })
-    }
-}
-
-/// Receiver Posting Key
-pub struct ReceiverPostingKey<C, L>
-where
-    C: Configuration,
-    L: ReceiverLedger<C> + ?Sized,
-{
-    /// UTXO Posting Key
-    utxo: L::ValidUtxo,
-
-    /// Encrypted Note
-    encrypted_note: EncryptedNote<C>,
-}
-
-impl<C, L> ReceiverPostingKey<C, L>
-where
-    C: Configuration,
-    L: ReceiverLedger<C> + ?Sized,
-{
-    /// Returns the ephemeral public key associated to `self`.
-    #[inline]
-    pub fn ephemeral_public_key(&self) -> &PublicKey<C> {
-        self.encrypted_note.ephemeral_public_key()
-    }
-
-    /// Extends proof public input with `self`.
-    #[inline]
-    pub fn extend_input(&self, input: &mut ProofInput<C>) {
-        C::ProofSystem::extend(input, self.utxo.as_ref());
-    }
-
-    /// Posts `self` to the receiver `ledger`.
-    #[inline]
-    pub fn post(self, super_key: &L::SuperPostingKey, ledger: &mut L) {
-        ledger.register(self.utxo, self.encrypted_note, super_key);
-    }
-
-    /// Posts all the of the [`ReceiverPostingKey`] in `iter` to the receiver `ledger`.
-    #[inline]
-    pub fn post_all<I>(iter: I, super_key: &L::SuperPostingKey, ledger: &mut L)
-    where
-        I: IntoIterator<Item = Self>,
-    {
-        ledger.register_all(
-            iter.into_iter().map(move |k| (k.utxo, k.encrypted_note)),
-            super_key,
-        )
-    }
 }
 
 /// Transfer
